+++ designer/client/src/components/toolbarComponents/toolbarButtons/ToolbarButtons.tsx
@@ -1,7 +1,7 @@
 import type { PropsOf } from "@emotion/react";
 import { styled } from "@mui/material";
 import type { PropsWithChildren } from "react";
-import React, { createContext } from "react";
+import React, { createContext, forwardRef } from "react";
 
 const ToolbarButtonsContainer = styled("div")(() => ({
     display: "flex",
@@ -14,6 +14,7 @@ export enum ButtonsVariant {
     small = "small",
     label = "label",
     horizontal = "horizontal",
+    menu = "menu",
 }
 
 export const ToolbarButtonsContext = createContext<{ variant: ButtonsVariant }>({ variant: ButtonsVariant.label });
@@ -22,10 +23,10 @@ type Props = {
     variant?: ButtonsVariant;
 } & PropsOf<typeof ToolbarButtonsContainer>;
 
-export function ToolbarButtons({ variant, ...props }: PropsWithChildren<Props>): JSX.Element {
+export const ToolbarButtons = forwardRef<HTMLDivElement, PropsWithChildren<Props>>(function ToolbarButtons({ variant, ...props }, ref) {
     return (
         <ToolbarButtonsContext.Provider value={{ variant: variant || ButtonsVariant.label }}>
-            <ToolbarButtonsContainer {...props} />
+            <ToolbarButtonsContainer {...props} ref={ref} />
         </ToolbarButtonsContext.Provider>
     );
-}
+});